Rename packages under java/main from org.jboss to org.jboss.set
