+++ src/main/java/org/jboss/set/config/GitHubRawUrl.java
@@ -1,4 +1,4 @@
-package org.jboss.config;
+package org.jboss.set.config;
 
 import org.eclipse.microprofile.config.spi.Converter;
