+++ src/main/java/org/jboss/set/processing/NewIssueCollector.java
@@ -1,19 +1,24 @@
-package org.jboss.processing;
+package org.jboss.set.processing;
 
 import com.atlassian.jira.rest.client.api.domain.Issue;
+import io.quarkus.arc.Arc;
 import io.quarkus.logging.Log;
 import org.apache.camel.component.jira.JiraEndpoint;
 import org.apache.camel.component.jira.consumer.NewIssuesConsumer;
-import org.jboss.processing.state.SingleIssueState;
+import org.jboss.set.draw.Lottery;
+import org.jboss.set.processing.state.EveryIssueState;
+import org.jboss.set.processing.state.SingleIssueState;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class StaleIssueCollector extends NewIssuesConsumer implements Executable {
+public class NewIssueCollector extends NewIssuesConsumer implements Executable {
 
     private static final State state = new State();
 
-    StaleIssueCollector(JiraEndpoint jiraEndpoint) {
+    private final Lottery lottery;
+
+    NewIssueCollector(JiraEndpoint jiraEndpoint) {
         super(jiraEndpoint, exchange -> {
             Issue issue = exchange.getIn().getBody(Issue.class);
             if (issue != null) {
@@ -21,6 +26,7 @@ public class StaleIssueCollector extends NewIssuesConsumer implements Executable
             }
         });
         state.issueStates.clear();
+        lottery = Arc.container().instance(Lottery.class).get();
     }
 
     private static final class State {
@@ -31,5 +37,7 @@ public class StaleIssueCollector extends NewIssuesConsumer implements Executable
     public void execute() throws Exception {
         int issuesCount = doPoll();
         Log.infof("Found number of issues %d", issuesCount);
+        EveryIssueState issues = new EveryIssueState(state.issueStates);
+        lottery.run(issues);
     }
 }
