+++ src/main/java/org/jboss/set/testing/JiraCommand.java
@@ -1,4 +1,4 @@
-package org.jboss.testing;
+package org.jboss.set.testing;
 
 import jakarta.inject.Qualifier;
