+++ src/main/resources/META-INF/services/org.eclipse.microprofile.config.spi.Converter
@@ -1,1 +1,1 @@
-org.jboss.config.GitHubRawUrl$GitHubRepoToRawUrlConverter
+org.jboss.set.config.GitHubRawUrl$GitHubRepoToRawUrlConverter
