+++ src/test/java/org/jboss/set/draw/MaxIssuesGlobalOptionTest.java
@@ -3,9 +3,8 @@ package org.jboss.set.draw;
 import com.atlassian.jira.rest.client.api.domain.Issue;
 import io.quarkus.mailer.Mail;
 import io.quarkus.test.junit.QuarkusTest;
-import org.jboss.config.LotteryConfig;
-import org.jboss.draw.Lottery;
-import org.jboss.query.IssueStatus;
+import org.jboss.set.config.LotteryConfig;
+import org.jboss.set.query.IssueStatus;
 import org.jboss.set.wrappers.IssueWrapper;
 import org.junit.jupiter.api.Test;
 import picocli.CommandLine;
